requirements.txt
flask
torch
torchvision
tqdm
scipy
ml-collections
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
